+++ src/infrastructure/database/models/order_model.py
@@ -7,7 +7,7 @@ orders_table = sqlalchemy.Table(
     metadata,
     sqlalchemy.Column("id", sqlalchemy.String, primary_key=True),
     sqlalchemy.Column("status", sqlalchemy.String, nullable=False),
-    sqlalchemy.Column("total_price", sqlalchemy.Float, nullable=False),
+    sqlalchemy.Column("total_price", sqlalchemy.Numeric(10, 2), nullable=False),
     sqlalchemy.Column(
         "created_at",
         sqlalchemy.DateTime,
